+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## vUNRELEASED
+
+#### Note for packagers
+
+Files generated from sources are no longer checked-in the git repository. Instead, packagers should build them:
+
+- Bash completion file, generated by `make bash-completion`
+- Man pages, generated by `make manpages`
+
 ## v1.3.0
 
 This release includes a number of new features and bugfixes like the long-awaited propagation of apps' exit status.
