+++ rkt/bash_completion_gen.go
@@ -1,0 +1,32 @@
+// Copyright 2016 The rkt Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// This file is not part of the rkt binary. It is only executed manually via
+// the makefile.
+//+build ignore
+
+package main
+
+import (
+	"fmt"
+	"os"
+)
+
+func main() {
+	err := cmdRkt.GenBashCompletionFile("dist/bash_completion/rkt.bash")
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Failed to generate the bash completion file: %v\n", err)
+		os.Exit(1)
+	}
+}
