+++ rkt/manpages_gen.go
@@ -1,0 +1,39 @@
+// Copyright 2016 The rkt Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// This file is not part of the rkt binary. It is only executed manually via
+// the makefile.
+//+build ignore
+
+package main
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/spf13/cobra/doc"
+)
+
+func main() {
+	header := &doc.GenManHeader{
+		Title:   "rkt - App Container runtime",
+		Section: "1",
+	}
+
+	err := doc.GenManTree(cmdRkt, header, "dist/manpages")
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Failed to generate the man pages: %v\n", err)
+		os.Exit(1)
+	}
+}
