+++ rkt/rkt.mk
@@ -29,6 +29,14 @@ $(BGB_BINARY): $(MK_PATH) | $(BINDIR)
 
 include makelib/build_go_bin.mk
 
+manpages: | $(GOPATH_TO_CREATE)/src/$(REPO_PATH)
+	mkdir -p dist/manpages/
+	ls rkt/*.go | grep -vE '_test.go|main.go|_gen.go' | xargs go run rkt/manpages_gen.go
+
+bash-completion: | $(GOPATH_TO_CREATE)/src/$(REPO_PATH)
+	mkdir -p dist/bash_completion/
+	ls rkt/*.go | grep -vE '_test.go|main.go|_gen.go' | xargs go run rkt/bash_completion_gen.go
+
 $(call undefine-namespaces,LOCAL)
 # RKT_STAMP deliberately not cleared
 # RKT_BINARY deliberately not cleared
